+++ sentry/examples/failure-demo.rs
@@ -1,3 +1,5 @@
+#![allow(deprecated)]
+
 use failure::Fail;
 use failure_ as failure;
 use sentry_failure::capture_error;