+++ sentry/src/lib.rs
@@ -49,7 +49,6 @@
 //!
 //! * `backtrace`: Enables backtrace support.
 //! * `contexts`: Enables capturing device, os, and rust contexts.
-//! * `failure`: Enables support for the `failure` crate.
 //! * `panic`: Enables support for capturing panics.
 //! * `transport`: Enables the default transport, which is currently `reqwest` with `native-tls`.
 //!
@@ -58,6 +57,7 @@
 //! * `anyhow`: Enables support for the `anyhow` crate.
 //! * `debug-images`: Attaches a list of loaded libraries to events (currently only supported on unix).
 //! * `error-chain`: Enables support for the `error-chain` crate.
+//! * `failure`: Enables support for the `failure` crate.
 //! * `log`: Enables support for the `log` crate.
 //! * `env_logger`: Enables support for the `log` crate with additional `env_logger` support.
 //! * `slog`: Enables support for the `slog` crate.
@@ -113,6 +113,7 @@ pub mod integrations {
     pub use sentry_error_chain as error_chain;
     #[cfg(feature = "failure")]
     #[doc(inline)]
+    #[deprecated = "The `failure` integration is deprecated and will be removed in the future."]
     pub use sentry_failure as failure;
     #[cfg(feature = "log")]
     #[doc(inline)]